test: refactor connection test functions for clarity and consistency
