+++ taos-ws-sys/src/ws/mod.rs
@@ -517,7 +517,7 @@ impl ResultSetOperations for ResultSet {
 }
 
 #[cfg(test)]
-pub unsafe fn test_connect() -> *mut TAOS {
+pub unsafe fn test_ws_connect() -> *mut TAOS {
     let taos = taos_connect(
         c"localhost".as_ptr(),
         c"root".as_ptr(),
@@ -530,7 +530,7 @@ pub unsafe fn test_connect() -> *mut TAOS {
 }
 
 #[cfg(test)]
-pub unsafe fn test_exec<S: AsRef<str>>(taos: *mut TAOS, sql: S) {
+pub unsafe fn test_ws_exec<S: AsRef<str>>(taos: *mut TAOS, sql: S) {
     let sql = std::ffi::CString::new(sql.as_ref()).unwrap();
     let res = query::taos_query(taos, sql.as_ptr());
     if res.is_null() {
@@ -543,13 +543,13 @@ pub unsafe fn test_exec<S: AsRef<str>>(taos: *mut TAOS, sql: S) {
 }
 
 #[cfg(test)]
-pub unsafe fn test_exec_many<T, S>(taos: *mut TAOS, sqls: S)
+pub unsafe fn test_ws_exec_many<T, S>(taos: *mut TAOS, sqls: S)
 where
     T: AsRef<str>,
     S: IntoIterator<Item = T>,
 {
     for sql in sqls {
-        test_exec(taos, sql);
+        test_ws_exec(taos, sql);
     }
 }
 
@@ -592,28 +592,6 @@ mod tests {
 
             let taos = taos_connect(ptr::null(), ptr::null(), ptr::null(), invalid_utf8_ptr, 0);
             assert!(taos.is_null());
-        }
-    }
-
-    #[test]
-    fn test_taos_connect_unable_to_establish_connection() {
-        unsafe {
-            let taos = taos_connect(
-                c"invalid_host".as_ptr(),
-                c"root".as_ptr(),
-                c"taosdata".as_ptr(),
-                ptr::null(),
-                6041,
-            );
-            assert!(taos.is_null());
-
-            let code = taos_errno(ptr::null_mut());
-            let errstr = taos_errstr(ptr::null_mut());
-            assert_eq!(Code::from(code), Code::new(0x000B));
-            assert_eq!(
-                CStr::from_ptr(errstr).to_str().unwrap(),
-                "Unable to establish connection"
-            );
         }
     }
